requirements.txt
flask
torch
torchvision
ultralytics
opencv-python
pillow
numpy
